# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/me": {
      "get": {
        "operationId": "UsersController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
...
    "/me/notifications": {
      "get": {
        "operationId": "NotificationsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 500,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/me/notifications/unread-count": {
      "get": {
        "operationId": "NotificationsController_unread",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/me/notifications/read-all": {
      "post": {
        "operationId": "NotificationsController_readAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/me/notifications/{id}/read": {
      "post": {
        "operationId": "NotificationsController_read",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/me/notification-preferences": {
      "get": {
        "operationId": "NotificationsController_preferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      },
      "put": {
        "operationId": "NotificationsController_setPreferences",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
    "/me/devices": {
      "post": {
        "operationId": "NotificationsController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      },
      "delete": {
        "operationId": "NotificationsController_unregister",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "THY API",
    "description": "API de l'écosystème THY. Authentification : jeton d'accès (Bearer) obtenu par OTP (POST /auth/otp/request puis /auth/otp/verify). Erreurs : { error: { code, message, details?, requestId } }. Les routes « à plat » de THY Business (ex. /products) visent l'entreprise active du jeton (POST /businesses/{id}/activate).",
    "version": "1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "locale": {
            "type": "string",
            "enum": [
              "fr",
              "en"
            ]
          },
          "timezone": {
            "type": "string",
            "maxLength": 64
          }
        }
      },
...
      "SetPreferencesDto": {
        "type": "object",
        "properties": {
          "push": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            }
          }
        },
        "required": [
          "push"
        ]
      },
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 20,
            "maxLength": 4096
          },
          "platform": {
            "type": "string",
            "enum": [
              "android",
              "ios",
              "web"
            ]
          }
        },
        "required": [
          "token",
          "platform"
        ]
      },
      "UnregisterDeviceDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 20,
            "maxLength": 4096
          }
        },
        "required": [
          "token"
        ]
      },
...
    }
  }
}